Clean up and break out hook
