+++ src/js/EpisodeTranscriptCell.tsx
@@ -1,9 +1,7 @@
-import { useContext } from 'react';
 import styled from 'styled-components';
 import type { Word } from './types';
 import { TimePrefixedWord } from './TimePrefixedWord';
-import { AudioContext } from './audio/AudioContext';
-import { Colors } from './constants';
+import { useHighlightWords } from './transcriptHooks';
 
 const makeKey = (w: Word) => {
   return `${w.startTime}-${w.endTime}-${w.word}-${w.probability}`;
@@ -25,37 +23,19 @@ const StyledTd = styled.td`
   }
 `;
 
-const punc = ['.', '?', '!'];
+interface EpisodeTranscriptCellProps {
+  words: Word[];
+  episode: number;
+}
 
 export const EpisodeTranscriptCell = ({
   words,
   episode,
-}: {
-  words: Word[];
-  episode: number;
-}) => {
-  const { currentTime, playingEpisode } = useContext(AudioContext);
-  const currentWordIndex = words.findIndex((word, i, list) => {
-    return (
-      currentTime >= word.startTime &&
-      (currentTime <= list[i + 1]?.endTime ?? Infinity)
-    );
+}: EpisodeTranscriptCellProps) => {
+  const shouldHighlight = useHighlightWords({
+    words,
+    episode,
   });
-
-  const endOfSentenceIndex =
-    words
-      .slice(currentWordIndex)
-      .findIndex(w => punc.some(p => w.word.endsWith(p))) + currentWordIndex;
-
-  const maybeHighlight = (i: number) => {
-    if (
-      episode === playingEpisode &&
-      i >= currentWordIndex &&
-      i <= endOfSentenceIndex
-    ) {
-      return { background: Colors.lightPurple };
-    }
-  };
 
   return (
     <StyledTd>
@@ -65,7 +45,7 @@ export const EpisodeTranscriptCell = ({
             key={makeKey(word)}
             $timestamp={word.startTime}
             $showPrefix={i > 0 && i % 200 === 0}
-            style={maybeHighlight(i)}
+            style={shouldHighlight(i)}
           >
             {word.word}
           </TimePrefixedWord>
